Move urllib3 subclasses into separate package
