+++ b2sdk/requests/NOTICE
@@ -6,13 +6,3 @@ Changes made to the original source:
 requests.models.Response.iter_content has been overridden to pass `decode_content=False` argument to `self.raw.stream`
 in order to NOT decompress data based on Content-Encoding header
 requests.adapters.HTTPAdapter has been overriden to use patched Urllib3 connection pools
-
-
-Urllib3
-Copyright 2008-2011 Andrey Petrov and contributors.
-
-Copyright 2023 Backblaze Inc.
-Changes made to the original source:
-The following classes have been overriden for support HTTP 100-continue:
-urllib3.connection.HTTPConnection, urllib3.connection.VerifiedHTTPSConnection,
-urllib3.connectionpool.HTTPConnectionPool, urllib3.connectionpool.HTTPSConnectionPool
