+++ b2sdk/urllib3/_connection.py
