+++ b2sdk/urllib3/included_source_meta.py
@@ -10,16 +10,16 @@
 from b2sdk.included_sources import IncludedSourceMeta, add_included_source
 
 included_source_meta = IncludedSourceMeta(
-    'requests', 'Included in a revised form', {
+    'urllib3', 'Included in a revised form', {
         'NOTICE':
-            """Requests
-Copyright 2019 Kenneth Reitz
+            """Urllib3
+Copyright 2008-2011 Andrey Petrov and contributors.
 
-Copyright 2021 Backblaze Inc.
+Copyright 2023 Backblaze Inc.
 Changes made to the original source:
-requests.models.Response.iter_content has been overridden to pass `decode_content=False` argument to `self.raw.stream`
-in order to NOT decompress data based on Content-Encoding header
-requests.adapters.HTTPAdapter has been overriden to use patched Urllib3 connection pools"""
+The following classes have been overriden for support HTTP 100-continue:
+urllib3.connection.HTTPConnection, urllib3.connection.VerifiedHTTPSConnection,
+urllib3.connectionpool.HTTPConnectionPool, urllib3.connectionpool.HTTPSConnectionPool"""
     }
 )
 add_included_source(included_source_meta)
